Consolidate credentials dir logic
