+++ credentials_nix.go
@@ -24,10 +24,6 @@ func sourceHelpString(credentialFile string, clusterName string, shell string) s
 	return s
 }
 
-const defaultDotDir = ".carina"
-const defaultNonDotDir = "carina"
-const xdgDataHomeEnvVar = "XDG_DATA_HOME"
-
 func userHomeDir() (string, error) {
 	if os.Getenv("HOME") != "" {
 		return os.Getenv("HOME"), nil
@@ -37,25 +33,4 @@ func userHomeDir() (string, error) {
 		return "", err
 	}
 	return currentUser.HomeDir, nil
-}
-
-// CarinaCredentialsBaseDir get the current base directory for carina credentials
-func CarinaCredentialsBaseDir() (string, error) {
-	if os.Getenv(CarinaHomeDirEnvVar) != "" {
-		return os.Getenv(CarinaHomeDirEnvVar), nil
-	}
-	if os.Getenv(CredentialsBaseDirEnvVar) != "" {
-		return os.Getenv(CredentialsBaseDirEnvVar), nil
-	}
-
-	// Support XDG
-	if os.Getenv(xdgDataHomeEnvVar) != "" {
-		return filepath.Join(os.Getenv(xdgDataHomeEnvVar), defaultNonDotDir), nil
-	}
-
-	homeDir, err := userHomeDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(homeDir, defaultDotDir), nil
 }
